+++ packages/runner/sync-gmail/repository/organization_repository.go
@@ -11,7 +11,7 @@ import (
 
 type OrganizationRepository interface {
 	GetOrganizationWithDomain(ctx context.Context, tx neo4j.ManagedTransaction, tenant, domainId string) (*dbtype.Node, error)
-	CreateOrganization(ctx context.Context, tx neo4j.ManagedTransaction, tenant, name, relationship, stage, source, sourceOfTruth, appSource string, date time.Time, hide bool) (*dbtype.Node, error)
+	CreateOrganization(ctx context.Context, tx neo4j.ManagedTransaction, tenant, name, relationship, stage, leadSource, source, sourceOfTruth, appSource string, date time.Time, hide bool) (*dbtype.Node, error)
 	LinkDomainToOrganization(ctx context.Context, tx neo4j.ManagedTransaction, tenant, domainName, organizationId string) error
 }
 
@@ -45,7 +45,7 @@ func (r *organizationRepository) GetOrganizationWithDomain(ctx context.Context, 
 	}
 }
 
-func (r *organizationRepository) CreateOrganization(ctx context.Context, tx neo4j.ManagedTransaction, tenant, name, relationship, stage, source, sourceOfTruth, appSource string, date time.Time, hide bool) (*dbtype.Node, error) {
+func (r *organizationRepository) CreateOrganization(ctx context.Context, tx neo4j.ManagedTransaction, tenant, name, relationship, stage, leadSource, source, sourceOfTruth, appSource string, date time.Time, hide bool) (*dbtype.Node, error) {
 	query := "MATCH (t:Tenant {name:$tenant}) " +
 		" MERGE (t)<-[:ORGANIZATION_BELONGS_TO_TENANT]-(org:Organization {id:randomUUID()}) " +
 		" ON CREATE SET org.createdAt=$now, " +
@@ -57,6 +57,7 @@ func (r *organizationRepository) CreateOrganization(ctx context.Context, tx neo4
 		"				org.hide=$hide, " +
 		"				org.relationship=$relationship, " +
 		"				org.stage=$stage, " +
+		"				org.leadSource=$leadSource, " +
 		"				org.isCustomer=$isCustomer, " +
 		"				org.onboardingStatus=$onboardingStatus, " +
 		"				org:%s " +
@@ -74,6 +75,7 @@ func (r *organizationRepository) CreateOrganization(ctx context.Context, tx neo4
 			"isCustomer":       false,
 			"relationship":     relationship,
 			"stage":            stage,
+			"leadSource":       leadSource,
 			"onboardingStatus": "NOT_APPLICABLE",
 		})
 	if err != nil {
