+++ packages/server/customer-os-api-sdk/graph/model/models_gen.go
@@ -2035,6 +2035,7 @@ type Organization struct {
 	YearFounded              *int64                        `json:"yearFounded,omitempty"`
 	Stage                    *OrganizationStage            `json:"stage,omitempty"`
 	Relationship             *OrganizationRelationship     `json:"relationship,omitempty"`
+	LeadSource               *string                       `json:"leadSource,omitempty"`
 	Hide                     bool                          `json:"hide"`
 	Contacts                 *ContactsPage                 `json:"contacts"`
 	JobRoles                 []*JobRole                    `json:"jobRoles"`
@@ -2114,6 +2115,7 @@ type OrganizationInput struct {
 	AppSource          *string                   `json:"appSource,omitempty"`
 	Stage              *OrganizationStage        `json:"stage,omitempty"`
 	Relationship       *OrganizationRelationship `json:"relationship,omitempty"`
+	LeadSource         *string                   `json:"leadSource,omitempty"`
 	// Deprecated
 	FieldSets []*FieldSetInput `json:"fieldSets,omitempty"`
 	// Deprecated
