+++ packages/server/customer-os-neo4j-repository/entity/organization.go
@@ -43,6 +43,7 @@ type OrganizationEntity struct {
 	LogoUrl            string
 	Relationship       enum.OrganizationRelationship `neo4jDb:"property:relationship;lookupName:RELATIONSHIP;supportCaseSensitive:false"`
 	Stage              enum.OrganizationStage        `neo4jDb:"property:stage;lookupName:STAGE;supportCaseSensitive:false"`
+	LeadSource         string                        `neo4jDb:"property:leadSource;lookupName:LEAD_SOURCE;supportCaseSensitive:true"`
 
 	LinkedOrganizationType *string
